Added files for controllers and routes
These are not complete and there are likely going to be more.
Started adding the routes that will be added to index file to allow
for a more organized project.
Controllers have nothing inside but Todo Controller has a link to a
resource that shows what I would like to do for the controllers.
added middleware to a dedicated file but may add a folder to seperate it further
although this isn't necassary at the moment.