added list lookup and fixed a bug with delete lists procedure
there was a bug that caused the query to not run.
mysql cant run subqueries on tables that are about to be updated.
if you nest it one deeper and make it a virtual table it will work.

also added lookup to help with this.
allows you to see data from Users, Invite_List, and Lists tables.