adds the begining of a DB factory class
This class is necassary to manage a connection pool for MYSQL.
There have been very noticable performance issues due to not properly managing connections.
to mitigate this a connection pool with up to 100 connections has been created.

Don't add this to any of the queries yet.
A Query manager class will be made.
This class will not be a singleton and its job will be to request and use connections.
That way we can simply create a new instance of the manager and query it.
it will open and close connections properly so we cannot make that mistake