Changes routes, adds permissions and lists to api
Adds permissions and lists to api by allowing you to see what
permissions are on what list, and being able to change whether or not
people can contribute or only view.

Changes the routes to allow the front end to have a similar routing
system. This will allow someone to go to a specific address like
/accounts as if it were bookmarked. If I need to use data related to
accounts then I can use the api/accounts